refactor: token type 이동
