+++ src/libs/recoil/states/token.ts
@@ -1,7 +1,7 @@
 import {atom} from 'recoil';
 import {RecoilStateKeys} from 'libs/recoil/constants/keys';
 import {localStorageEffect} from 'libs/recoil/utils';
-import {TokenType} from 'libs/recoil/types';
+import {TokenType} from 'types/token';
 
 export const tokenState = atom<TokenType>({
   key: RecoilStateKeys.Token,
